Rename the Unlicense file for clarity
To be perfectly clear: this implementation of true is NOT PROTECTED BY
COPYRIGHT. It is clearly and explicitly released into the public domain.

Placing information about this in a file called "LICENSE.txt" makes no
sense. There is NO COPYRIGHT PROTECTION, so it is impossible for me to
grant any sort of "license" to this software.

Per the recommendation of unlicense.org, the file has been moved to a
more appropriate location and references have been updated as necessary.